+++ android/app/src/test/java/org/smartregister/fhircore/activity/LoginActivityTest.kt
@@ -19,7 +19,9 @@ package org.smartregister.fhircore.activity
 import android.app.Activity
 import android.content.Intent
 import android.os.Looper.getMainLooper
+import android.view.View
 import android.widget.Button
+import android.widget.ProgressBar
 import androidx.test.core.app.ApplicationProvider
 import org.junit.Assert
 import org.junit.Before
@@ -137,6 +139,18 @@ class LoginActivityTest : ActivityRobolectricTest() {
     val actualIntent =
       shadowOf(ApplicationProvider.getApplicationContext<FhirApplication>()).nextStartedActivity
     Assert.assertEquals(expectedIntent.component, actualIntent.component)
+  }
+
+  @Test
+  fun testVerifyProgressBarVisibilityWhenShowInProgressIconUpdate() {
+    val showProgressIcon = loginActivity.findViewById<ProgressBar>(R.id.progressIcon)
+    Assert.assertEquals(View.INVISIBLE, showProgressIcon.visibility)
+
+    loginActivity.viewModel.showProgressIcon.value = true
+    Assert.assertEquals(View.VISIBLE, showProgressIcon.visibility)
+
+    loginActivity.viewModel.showProgressIcon.value = false
+    Assert.assertEquals(View.INVISIBLE, showProgressIcon.visibility)
   }
 
   override fun getActivity(): Activity {